more service update and modularized
