+++ src/services/ledService.h
@@ -11,6 +11,15 @@ public:
     LedService(/* args */);
     ~LedService();
 
+    void InitLeds()
+    {
+        pinMode(PIN_RED, OUTPUT);
+        pinMode(PIN_GREEN, OUTPUT);
+        pinMode(PIN_BLUE, OUTPUT);
+        pinMode(LEDINBUILT, OUTPUT);
+        // Serial.begin(921600);
+    }
+
     void setLedColor(int R, int G, int B)
     {
         analogWrite(PIN_RED, R);
@@ -26,6 +35,5 @@ LedService::LedService(/* args */)
 LedService::~LedService()
 {
 }
-
 
 #endif
